Primarily a documentation and test fix commit.  Also added an easier dynamic library creation.
Prior to this commit, the testing of static library building was not actually testing the static library but was doing the usual cake job of compiling everything in the one go.  This commit fixes that issue.

Since I was working on library creation I added a command line switch, --dynamic-library, to make dynamic library creation as easy as static library creation. Added a test for dynamic library creation too.

The documentation about the /etc/cake.conf variables has been updated to capture the fact that a variant must now always be specified.  Also added the dynamic library creation to the doco.